Update dependencies; fix failing test.
One of the RSA interop tests was intermittently failing. It seems like
OpenSSL would sometimes exit before the write was flushed so no output
would be written to stdout, but it would still exit with code 0. Instead
of using `stdin.end(input)` as a shortcut for `stdin.write(input);
stdin.end();`, this patch changes the test to use the form
`stdin.write(input, callback)` and calls `stdin.end()` only after we're
sure the write has been fully flushed. This seems to have fixed the
problem: while before I was getting errors on about half or more of the
test runs, I was able to run the test suite 20 times without error.